Cracked the game timer value!
I noticed that the difference between values for 0x45e40 after an hour was
consistently 4,096, and dividing the value by that constant got me the precise
number of hours, with a remainder.

I already knew there was a link between 64 and the number of seconds, so on
a chance I divided the remainder by 64 got the exact number of minutes, and
the remainder was the exact integer number fo seconds.