Begin reworking compare & monitor tools
The Comparator.monitor_gen() method will eventually replace .monitor().
It is a simple generator, and all the work of driving it and producing
output is in the command line app.

Still todo: wire in the recording function and offload that
functionality out of the Comparator class and into the app.