Replace decryption/encryption functions
The new algorithms for decrypting (and encrypting) save data are based
on a gist published by lincoln-lm, which appears to implement the actual
algorithm used by the game. This means that the decrypt command can
decrypt without needing a key extracted from the main save file. This
allows the avatarmakedata and systemdata files to be decrypted.

As a consequence of the change, the encrypt and decrypt commands now
require a specific flag to work with Definitive Edition files (due to
byte-order differences)